renaming directory tree item and minor changes
