+++ src/azureStorageExplorer/FileStatImpl.ts
@@ -1,0 +1,27 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import * as vscode from 'vscode';
+
+export class FileStatImpl implements vscode.FileStat {
+    // tslint:disable-next-line: no-reserved-keywords
+    type: vscode.FileType;
+    ctime: number;
+    mtime: number;
+    size: number;
+
+    constructor(
+        // tslint:disable-next-line: no-reserved-keywords
+        type: vscode.FileType,
+        ctime: number,
+        mtime: number,
+        size: number
+    ) {
+        this.type = type;
+        this.ctime = ctime;
+        this.mtime = mtime;
+        this.size = size;
+    }
+}
